requirements.txt
alembic
click
Flask
Flask-Admin
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
itsdangerous
Jinja2
Mako
MarkupSafe
psycopg2-binary
python-dotenv
SQLAlchemy
Werkzeug
WTForms
email-validator